[Refactor] web 패키지 제거 및 통합
