+++ src/main/java/com/inq/wishhair/wesharewishhair/hairstyle/controller/dto/response/PagedHairStyleResponse.java
@@ -1,0 +1,13 @@
+package com.inq.wishhair.wesharewishhair.hairstyle.controller.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class PagedHairStyleResponse<T> {
+
+    private T result;
+
+    private int contentSize;
+}
