+++ src/main/java/com/inq/wishhair/wesharewishhair/wishlist/domain/WishList.java
@@ -1,0 +1,39 @@
+package com.inq.wishhair.wesharewishhair.wishlist.domain;
+
+import com.inq.wishhair.wesharewishhair.auditing.BaseEntity;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
+import com.inq.wishhair.wesharewishhair.user.domain.User;
+import jakarta.persistence.*;
+import lombok.Builder;
+import lombok.Getter;
+
+@Entity
+@Getter
+public class WishList extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "hair_style_id")
+    private HairStyle hairStyle;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    //==생성 메서드==//
+    @Builder
+    public WishList() {
+    }
+
+    //==편의 메서드==//
+    public void registerHairStyle(HairStyle hairStyle) {
+        this.hairStyle = hairStyle;
+    }
+
+    public void registerUser(User user) {
+        this.user = user;
+    }
+}
