+++ src/main/java/com/inq/wishhair/wesharewishhair/wishlist/service/WishListService.java
@@ -4,8 +4,8 @@ import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
 import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyleRepository;
 import com.inq.wishhair.wesharewishhair.user.domain.User;
 import com.inq.wishhair.wesharewishhair.user.domain.UserRepository;
-import com.inq.wishhair.wesharewishhair.wishlist.WishList;
-import com.inq.wishhair.wesharewishhair.wishlist.repository.WishListRepository;
+import com.inq.wishhair.wesharewishhair.wishlist.domain.WishList;
+import com.inq.wishhair.wesharewishhair.wishlist.domain.WishListRepository;
 import com.inq.wishhair.wesharewishhair.global.exception.ErrorCode;
 import com.inq.wishhair.wesharewishhair.global.exception.WishHairException;
 import lombok.RequiredArgsConstructor;
