+++ src/test/java/com/inq/wishhair/wesharewishhair/global/base/ControllerTest.java
@@ -5,9 +5,9 @@ import com.inq.wishhair.wesharewishhair.hairstyle.service.HairStyleService;
 import com.inq.wishhair.wesharewishhair.domain.login.LoginService;
 import com.inq.wishhair.wesharewishhair.point.service.PointHistoryService;
 import com.inq.wishhair.wesharewishhair.user.service.UserService;
-import com.inq.wishhair.wesharewishhair.web.hairstyle.HairStyleController;
+import com.inq.wishhair.wesharewishhair.hairstyle.controller.HairStyleController;
 import com.inq.wishhair.wesharewishhair.web.login.LoginController;
-import com.inq.wishhair.wesharewishhair.web.user.UserController;
+import com.inq.wishhair.wesharewishhair.user.controller.UserController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
