+++ src/test/java/com/inq/wishhair/wesharewishhair/user/controller/UserControllerTest.java
@@ -2,7 +2,7 @@ package com.inq.wishhair.wesharewishhair.user.controller;
 
 import com.inq.wishhair.wesharewishhair.global.base.ControllerTest;
 import com.inq.wishhair.wesharewishhair.user.controller.utils.UserCreateRequestUtils;
-import com.inq.wishhair.wesharewishhair.web.user.dto.request.UserCreateRequest;
+import com.inq.wishhair.wesharewishhair.user.controller.dto.request.UserCreateRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
